Renamed 'operators.js' file and moved to a new directory
The 'operators.js' file was moved from '../js-concepts/types' to a new directory at '../js-concepts/operators'. This also inclues the renaming of the file to 'index.js' for improved file structure and organization, with the intent of making its purpose clearer - to provide an understanding of different JavaScript operators' usage.