Add 'coercion.js' to explain JavaScript coercion
A new file, 'coercion.js', has been added to the 'js-concepts/operators' directory. This file is dedicated to giving a detailed explanation of JavaScript coercion. It elucidates the process of converting a value from one type to another with provided code examples, assisting with the comprehension of dynamic data type conversion in JavaScript.